altered action listeners
