+++ src/edu/sjsu/view/LoginViewer.java
@@ -22,6 +22,7 @@ public class LoginViewer extends BankViewer {
 
         addLoginPanel();
         addRegisterPanel();
+        
         this.setVisible(true);
     }
